Various Changes 5.875

+++ TweakUtility/TweakPages/UncategorizedPage.cs
@@ -40,6 +40,7 @@ namespace TweakUtility.TweakPages
             set => RegistryHelper.SetValue(@"HKLM\SYSTEM\CurrentControlSet\Control\Terminal Server\WinStations\RDP-Tcp\PortNumber", value);
         }
 
+        [Browsable(true)]
         [Category("Windows Media Player")]
         [DisplayName("Install Deskband")]
         public void InstallWMPDeskBand(ProgressIndicator indicator)
@@ -48,26 +49,34 @@ namespace TweakUtility.TweakPages
 
             using (var client = new WebClient())
             {
-                indicator.SetProgress(1, "Downloading 32-bit deskband...");
-                client.DownloadFile("https://raw.githubusercontent.com/Craftplacer/TweakUtility/master/Optional/wmpband/32.dll", "32.dll");
-
-                indicator.SetProgress(2, "Downloading 64-bit deskband...");
-                client.DownloadFile("https://raw.githubusercontent.com/Craftplacer/TweakUtility/master/Optional/wmpband/64.dll", "64.dll");
-
-                indicator.SetProgress(3, "Moving files...");
                 string x86 = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86), "Windows Media Player", "wmpband.dll");
-                File.Move("32.dll", x86);
-
                 string x64 = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles), "Windows Media Player", "wmpband.dll");
-                File.Move("64.dll", x64);
+
+                if (!File.Exists(x86))
+                {
+                    indicator.SetProgress(1, "Downloading 32-bit deskband...");
+                    client.DownloadFile("https://raw.githubusercontent.com/Craftplacer/TweakUtility/master/Optional/wmpband/32.dll", "32.dll");
+
+                    indicator.SetProgress(3, "Moving 32-bit deskband file...");
+                    File.Move("32.dll", x86);
+                }
+
+                if (!File.Exists(x64))
+                {
+                    indicator.SetProgress(2, "Downloading 64-bit deskband...");
+                    client.DownloadFile("https://raw.githubusercontent.com/Craftplacer/TweakUtility/master/Optional/wmpband/64.dll", "64.dll");
+
+                    indicator.SetProgress(3, "Moving 64-bit deskband file...");
+                    File.Move("64.dll", x64);
+                }
 
                 string regsvrPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.System), "regsvr32.exe");
 
                 indicator.SetProgress(4, "Registering 32-bit deskband...");
-                Process.Start(regsvrPath, x86).WaitForExit();
+                Process.Start(regsvrPath, $"\"{x86}\"").WaitForExit();
 
                 indicator.SetProgress(5, "Registering 64-bit deskband...");
-                Process.Start(regsvrPath, x64).WaitForExit();
+                Process.Start(regsvrPath, $"\"{x64}\"").WaitForExit();
             }
         }
